Subject ID touch-ups
Saved dbpaths and dbt into my (Erik's) Y:/ folder, but wrapped them in a
try block in case I change the folder structure. Also, subject IDs with
typos in the name are now corrected when they are read in from the
metadata, instead of before comparing them to the pathology excel sheet.
Also re-labelled subject 30 as "Practice" because it should not be
associated with a real subject ID. Now all subject IDs in dbt are
correct, with no typos.