Data Visualization Tools added
Several Data visualization tools have been added for use. Histograms and
boxplots which show more details of the data and an attempt at the
subject analysis are the major changes.

This also adds several other peoples packages, Liscences and such have
been kept and they are in seperate folders to show that difference.